requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
bottle==0.12.25
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chatglm-cpp @ file:///C:/Users/jophy/Downloads/chatglm_cpp-0.3.0-cp39-cp39-win_amd64.whl
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.0
orjson==3.9.10
packaging==23.2
pefile==2023.2.7
proxy-tools==0.1.0
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
python-dotenv==1.0.0
python-multipart==0.0.6
pythonnet==3.0.3
pywebview==4.4.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
sse-starlette==1.8.2
starlette==0.32.0.post1
tqdm==4.66.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
zipp==3.17.0
